+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog of QuickLinkJs:
+# Changelog of Linkutter:
 
 - v0.0.1 Alpha A-1 {master} {
     - Inital Commit, Added module files into the repository.
@@ -7,4 +7,14 @@
 - v0.0.1 Alpha A-1 Update0.1 {master} {
     - Added some features :+1: !
     - Fixed some bugs :bug: !
+}
+
+- v0.0.1 Alpha A-1 Update0.2 {master} {
+    - Added some features :+1: !
+        - New animations :100: !
+        - Improved navbar :rocket: !
+        - Improved user comfort !
+        - :new: NEW: custom context menu !
+        - :new: popup to display info's
+    - Fixed a lot bugs :bug: !
 }